requirements.txt
torch==2.4.0
torchvision==0.19.0
diffusers==0.25.0
transformers==4.39.3
tokenizers==0.15.2
huggingface_hub==0.19.4
accelerate
fastapi
uvicorn
pillow
numpy
